Changed request in SymptomsController to retrun symptoms by bodypartname

+++ src/main/java/bg/healthcheck/BIYD/Entities/Symptoms.java
@@ -17,7 +17,7 @@ public class Symptoms {
     private String name;
 
     @Column(name="body_part_id")
-    private Integer body_part;
+    private Long body_part;
 
     @ManyToMany(mappedBy = "symptoms")
     @JsonIgnore
@@ -26,7 +26,7 @@ public class Symptoms {
     public Symptoms() {
     }
 
-    public Symptoms(Long id, String name, Integer body_part, Set<Illnesses> illnesses) {
+    public Symptoms(Long id, String name, Long body_part, Set<Illnesses> illnesses) {
         this.id = id;
         this.name = name;
         this.body_part = body_part;
@@ -45,11 +45,11 @@ public class Symptoms {
         this.name = name;
     }
 
-    public Integer getBody_part() {
+    public Long getBody_part() {
         return body_part;
     }
 
-    public void setBody_part(Integer body_part) {
+    public void setBody_part(Long body_part) {
         this.body_part = body_part;
     }
